ref: move getNewValueForFieldOp to validation. Added onInit
