+++ packages/core/modules/import/index.js
@@ -1,3 +1,9 @@
-export * from "./tree";
-export * from "./jsonLogic";
-export * from "./spel";
+export {
+  getTree, loadTree, isValidTree,
+  // deprecated
+  checkTree,
+  // candidates for moving to utils
+  isImmutableTree, isTree, isJsonLogic, jsToImmutable
+} from "./tree";
+export { loadFromJsonLogic, _loadFromJsonLogic } from "./jsonLogic";
+export { loadFromSpel, _loadFromSpel } from "./spel";
