+++ packages/core/modules/utils/index.js
@@ -1,7 +1,5 @@
 export {default as clone} from "clone";
 export {default as moment} from "moment";
-export {validateTree, sanitizeTree} from "./validation";
-export {default as uuid} from "./uuid";
 export * as ConfigUtils from "./configUtils";
 export * as RuleUtils from "./ruleUtils";
 export * as FuncUtils from "./funcUtils";
@@ -12,5 +10,9 @@ export * as ListUtils from "./listValues";
 export * as Autocomplete from "./autocomplete";
 export * as Validation from "./validation";
 export * as OtherUtils from "./stuff";
+
+// for backward compatibility:
+export {validateTree, sanitizeTree} from "./validation";
+export {default as uuid} from "./uuid";
 export {getSwitchValues} from "./treeUtils";
 export {compressConfig, decompressConfig} from "./configSerialize";
